+++ src/core/create-sample-config.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs';
 import chalk from "chalk";
 import { TypeSharpConfig } from "../types/index.js";
+import { logger } from '../helpers/logger.js';
 
 
 /**
@@ -42,7 +43,8 @@ export const createSampleConfig = (format: 'ts' | 'js' | 'json'): void => {
     const allConfigFiles = ['typesharp.config.ts', 'typesharp.config.js', 'typesharp.config.json'];
     const existingConfig = allConfigFiles.find(f => fs.existsSync(f));
     if (existingConfig) {
-        console.log(chalk.yellow.bold('❗ Warning:'), chalk.white(`${existingConfig} already exists. Skipping creation.`));
+        console.log('')
+        logger.warn('createSampleConfig', `${existingConfig} already exists\n`)
         return;
     }